add integration tests, fix system tests
