+++ packages/zosjobs/__tests__/__integration__/cli/submit/data-set/__scripts__/submit_syntax_wait_active_and_directory.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+zowe zos-jobs submit data-set "DATA.SET" --wait-for-active --directory "./thisshouldgetasyntaxerror"
+exit $?
