+++ packages/zosjobs/__tests__/__integration__/cli/submit/data-set/__scripts__/submit_syntax_wait_active_and_vasc.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+# should get a syntax error
+zowe zos-jobs submit data-set "DATA.SET" --wait-for-active --vasc
+exit $?
