+++ packages/zosjobs/__tests__/__integration__/cli/submit/data-set/cli.zos-jobs.submit.data-set.integration.test.ts
@@ -59,5 +59,19 @@ describe("zos-jobs submit data-set command", () => {
             expect(response.stdout.toString()).toBe("");
             expect(response.stderr.toString()).toMatchSnapshot();
         });
+
+        it("should occur if both --wait-for-active and --directory is specified", async () => {
+            const response = runCliScript(__dirname + "/__scripts__/submit_syntax_wait_active_and_directory.sh", TEST_ENVIRONMENT);
+            expect(response.status).toBe(1);
+            expect(response.stdout.toString()).toBe("");
+            expect(response.stderr.toString()).toMatchSnapshot();
+        });
+
+        it("should occur if both --wait-for-active and --directory is specified", async () => {
+            const response = runCliScript(__dirname + "/__scripts__/submit_syntax_wait_active_and_vasc.sh", TEST_ENVIRONMENT);
+            expect(response.status).toBe(1);
+            expect(response.stdout.toString()).toBe("");
+            expect(response.stderr.toString()).toMatchSnapshot();
+        });
     });
 });
