+++ packages/zosjobs/__tests__/__integration__/cli/submit/local-file/__scripts__/submit_syntax_wait_active_and_directory.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+zowe zos-jobs submit local-file $1 --wait-for-active --directory "./syntaxerr"
+exit $?
