+++ packages/zosjobs/__tests__/__integration__/cli/submit/stdin/cli.zos-jobs.submit.stdin.integration.test.ts
@@ -13,47 +13,28 @@ import {ITestEnvironment} from "../../../../../../../__tests__/__src__/environme
 import {TestEnvironment} from "../../../../../../../__tests__/__src__/environment/TestEnvironment";
 import {runCliScript} from "../../../../../../../__tests__/__src__/TestUtils";
 
+process.env.FORCE_COLOR = "0";
+
 // Test Environment populated in the beforeAll();
 let TEST_ENVIRONMENT: ITestEnvironment;
 
-describe("zos-jobs submit local-file command", () => {
+describe("zos-jobs submit command", () => {
     // Create the unique test environment
     beforeAll(async () => {
         TEST_ENVIRONMENT = await TestEnvironment.setUp({
-            testName: "zos_jobs_submit_local_file_command",
+            testName: "zos_jobs_submit_command",
             skipProperties: true
         });
     });
 
-    afterAll(async () => {
-        await TestEnvironment.cleanUp(TEST_ENVIRONMENT);
+    it("should display the help", async () => {
+        const response = runCliScript(__dirname + "/__scripts__/submit_help.sh", TEST_ENVIRONMENT);
+        expect(response.stderr.toString()).toBe("");
+        expect(response.status).toBe(0);
+        expect(response.stdout.toString()).toMatchSnapshot();
     });
 
-    it("should fail if the local file does not exist ", async () => {
-        const response = runCliScript(__dirname + "/__scripts__/submit_invalid_local_file.sh",
-            TEST_ENVIRONMENT);
-        expect(response.status).toBe(1);
-        expect(response.stderr.toString().toLowerCase()).toContain("error");
-        expect(response.stderr.toString().toLowerCase()).toContain("no such file");
-    });
-
-    describe("syntax errors", () => {
-        it("should occur if the local file name is missing", async () => {
-            const response = runCliScript(__dirname + "/__scripts__/submit_syntax_missing_file_name.sh", TEST_ENVIRONMENT);
-            expect(response.status).toBe(1);
-            expect(response.stdout.toString()).toBe("");
-            expect(response.stderr.toString()).toMatchSnapshot();
-        });
-
-        it("should occur if an extra unknown option is specified", async () => {
-
-            const response = runCliScript(__dirname + "/__scripts__/submit_syntax_invalid_parm.sh",
-                TEST_ENVIRONMENT, [__dirname + "/testFileOfLocalJCL.txt"]);
-            expect(response.status).toBe(1);
-            expect(response.stdout.toString()).toBe("");
-            expect(response.stderr.toString()).toMatchSnapshot();
-        });
-
+    describe("Syntax errors", async () => {
         it("should occur if --wfa and --directory are both specified", async () => {
 
             const response = runCliScript(__dirname + "/__scripts__/submit_syntax_wait_active_and_directory.sh",
