+++ packages/zosjobs/__tests__/__system__/cli/submit/data-set/__scripts__/submit_valid_data_set_wait.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 # pass the data set name as an argument to the script
-zowe zos-jobs submit data-set $1 -w
+zowe zos-jobs submit data-set $1 --wfo
 exit $?
