+++ packages/zosjobs/__tests__/__system__/cli/submit/local-file/__scripts__/submit_valid_local_file_wait.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 # pass the data set name as an argument to the script
-zowe zos-jobs submit local-file $1 --wait
+zowe zos-jobs submit local-file $1 --wait-for-output
 exit $?
