+++ packages/zosjobs/__tests__/__system__/cli/submit/stdin/__scripts__/submit_valid_stdin_wait.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 # pass the JCL as stdin to this script
-cat $1 | zowe zos-jobs submit stdin -w
+cat $1 | zowe zos-jobs submit stdin --wfo
 exit $?
